reminders-service/requirements.txt
supabase==2.10.0
httpx==0.27.2
APScheduler==3.10.4
python-dotenv==1.0.1
tenacity==9.0.0
phonenumbers==8.13.47
